重构基类文件，重构Repository
