+++ src/Senparc.Core/Models/VD/ErrorVD.Core.cs
@@ -1,7 +1,7 @@
 ﻿
 using System;
 
-namespace Senparc.Scf.Core.Models.VD
+namespace Senparc.Core.Models.VD
 {
     /// <summary>
     /// 全局所有的Error_ExceptionVD必须实现这个接口
