+++ src/Senparc.Core/Models/VD/HomeVD.cs
