+++ src/Senparc.Core/Models/VD/LoginVD.cs
