+++ src/Senparc.Scf.Repository/BaseRepoisitory/ClientRepositoryBase.cs
@@ -6,27 +6,29 @@ namespace Senparc.Scf.Repository
 {
     public interface IClientRepositoryBase<T> : IRepositoryBase<T> where T : EntityBase // global::System.Data.Objects.DataClasses.EntityObject, new()
     {
-        ISqlClientFinanceData DB { get; }
+        ISqlBaseFinanceData DB { get; }
     }
 
     public class ClientRepositoryBase<T> : RepositoryBase<T>, IClientRepositoryBase<T> where T : EntityBase // global::System.Data.Objects.DataClasses.EntityObject, new()
     {
-        public ISqlClientFinanceData DB
+        public ISqlBaseFinanceData DB
         {
             get
             {
-                return base.BaseDB as ISqlClientFinanceData;
+                return base.BaseDB; //as ISqlClientFinanceData;
             }
         }
 
         public ClientRepositoryBase() : this(null) { }
-        public ClientRepositoryBase(ISqlClientFinanceData db)
+        public ClientRepositoryBase(ISqlBaseFinanceData db)
         {
             //System.Web.HttpContext.Current.Response.Write("-"+this.GetType().Name + "<br />");
 
-            base.BaseDB = db ?? SenparcDI.GetService<ISqlClientFinanceData>();// ObjectFactory.GetInstance<ISqlClientFinanceData>();
+            base.BaseDB = db ?? SenparcDI.GetService<ISqlBaseFinanceData>();// ObjectFactory.GetInstance<ISqlClientFinanceData>();
 
-            _entitySetName = EntitySetKeys.Keys[typeof(T)];
+
+            var keys = EntitySetKeys.GetEntitySetKeys(db.GetType());
+            _entitySetName = keys[typeof(T)];
         }
